Comment submission

+++ backend/models/user-comments-manager.js
@@ -33,18 +33,21 @@ const USER_COMMENTS_FILE_PATH = path.join(__dirname, '../data/user-comments.json
 module.exports = class UserCommentsManager {
 
   static async recordComment(user, comment, antispamAnswer, isAdministrator = false) {
-    if (!this.testIfUserCommentsFileExists()) {
-      this.createUserCommentsFile();
-    }
-    if (!checkAntispamAnswer(antispamAnswer)) {
-      return { status: "error", message: "Mauvaise réponse à la question anti-spam" };
-    }
+    await this.createUserCommentsFileIfNotExists();
+
+    if (!checkAntispamAnswer(antispamAnswer)) return { status: "error", message: "Mauvaise réponse à la question anti-spam" };
+    if (!user.name || user.name.length > 50) return { status: "error", message: "Nom invalide" };
+    if (user.email && user.email.length > 100) return { status: "error", message: "Email invalide" };
+    if (!comment || comment.length > 1100) return { status: "error", message: "Commentaire invalide" };
+    if (!antispamAnswer.id || !antispamAnswer.answer) return { status: "error", message: "Réponse anti-spam invalide" };
     const userComments = JSON.parse(fs.readFileSync(USER_COMMENTS_FILE_PATH, 'utf8'));
     const newCommentId = userComments.reduce((acc, comment) => comment.id > acc ? comment.id : acc, 0) + 1;
-    const userComment = { id: newCommentId, user: user, comment: comment };
+    const DateTimeISO = new Date().toISOString();
+    const userComment = { id: newCommentId, user: user, comment: comment, date: DateTimeISO };
     if (isAdministrator) userComment.isAdministrator = true;
     userComments.push(userComment);
     fs.writeFileSync(USER_COMMENTS_FILE_PATH, JSON.stringify(userComments));
+    return { status: "ok" };
   }
 
   static async deleteUserComment(commentId) {
@@ -58,6 +61,7 @@ module.exports = class UserCommentsManager {
     }
     userComments.splice(index, 1);
     fs.writeFileSync(USER_COMMENTS_FILE_PATH, JSON.stringify(userComments));
+    return { status: "ok" };
   }
 
   static async getComments() {
@@ -71,12 +75,13 @@ module.exports = class UserCommentsManager {
     return userComments;
   }
 
-  static async createUserCommentsFile() {
-    fs.writeFileSync(USER_COMMENTS_FILE_PATH, JSON.stringify([]));
+  static async createUserCommentsFileIfNotExists() {
+    const existFile = fs.existsSync(USER_COMMENTS_FILE_PATH);
+    if (!existFile) fs.writeFileSync(USER_COMMENTS_FILE_PATH, JSON.stringify([]));
   }
 
   static async testIfUserCommentsFileExists() {
-    return fs.existsSync(USER_COMMENTS_FILE_PATH);
+    return
   }
 
   static getAntispamQuestion() {
@@ -89,9 +94,8 @@ module.exports = class UserCommentsManager {
 
 function checkAntispamAnswer(antispamAnswer) {
   const id = parseInt(antispamAnswer.id);
-  const answer = antispamAnswer.answer;
+  const answer = antispamAnswer.answer.trim().toLowerCase();
   const question = antispamQuestions.find(question => question.id === id);
-  answer = answer.trim().toLowerCase();
   const answerRegex = /^[a-z0-9]{1,20}$/;
   if (!answerRegex.test(answer)) {
     return false;
